make dispatch.py slimmer
